# backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Vinculación",
    "version": "1.0.0",
    "description": "Documentación automática de todos los endpoints del sistema de vinculación"
  },
  "host": "localhost:3005",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Token JWT en formato: Bearer {token}"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/login": {
      "post": {
        "tags": [
          "Login"
        ],
        "description": "Endpoint para autenticar usuarios",
        "parameters": [
          {
            "name": "x-request-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "deviceId": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "deviceInfo": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "415": {
            "description": "Unsupported Media Type"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/refresh": {
      "post": {
        "tags": [
          "Login"
        ],
        "description": "Obtener nuevo access token usando cookie HttpOnly de refresh y deviceId. Validación de metadatos del dispositivo.",
        "parameters": [
          {
            "name": "x-request-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "user-agent",
            "in": "header",
            "type": "string"
          },
          {
            "name": "cookie",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "415": {
            "description": "Unsupported Media Type"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/login/google": {
      "post": {
        "tags": [
          "Login"
        ],
        "description": "Iniciar sesión o registrarse con Google OAuth",
        "parameters": [
          {
            "name": "x-request-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "deviceId": {
                  "example": "any"
                },
                "id_token": {
                  "example": "any"
                },
                "deviceInfo": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "415": {
            "description": "Unsupported Media Type"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/logout": {
      "post": {
        "tags": [
          "Login"
        ],
        "description": "Endpoint para cerrar sesión",
        "parameters": [
          {
            "name": "x-request-id",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/me": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "x-request-id",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
...
    "/api/menus": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "x-request-id",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
...
    "/api/menus/{id_rol}": {
      "get": {
        "tags": [
          "Access"
        ],
        "description": "Obtiene los menús disponibles según el rol",
        "parameters": [
          {
            "name": "id_rol",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "x-request-id",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
...
    "/api/getnotifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "description": "Obtiene todas las notificaciones",
        "parameters": [
          {
            "name": "x-request-id",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/getnotificationbyid/{id}": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "description": "Obtiene una notificación por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "x-request-id",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/postnotification": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "description": "Crea una nueva notificación - Solo administradores",
        "parameters": [
          {
            "name": "x-request-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "example": "any"
                },
                "type": {
                  "example": "any"
                },
                "message": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/patchnotificationstatus/{id}": {
      "patch": {
        "tags": [
          "Notifications"
        ],
        "description": "Actualiza el estado de una notificación",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "x-request-id",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/deletenotificationbyid/{id}": {
      "delete": {
        "tags": [
          "Notifications"
        ],
        "description": "Elimina una notificación",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "x-request-id",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    }
  },
...
}